+++ app/Http/Controllers/RoleController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class RoleController extends Controller
 {
 
     public function index()
     {
-        $roles = Role::latest()->get();
+        $roles = Role::where('slug','!=','super-admin')->latest()->get();
         return view('backpanel.role-permission.roles', compact('roles'));
     }
 
@@ -24,7 +25,7 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         if($request->has('id')){
-            $role = Role::find($request->id);
+            $role = Role::where('slug','!=','super-admin')->where('id',$request->id)->first();
             $request->session()->flash('success', 'Role Updated successfully!');
         }else{
             $role = new Role();
@@ -39,6 +40,10 @@ class RoleController extends Controller
 
     public function edit($id)
     {
+        $count = Role::where('slug','super-admin')->where('id',$id)->count();
+        if($count > 0){
+            return redirect()->back();
+        }
         $role = Role::find($id);
         $permissions = Permission::get();
         return view('backpanel.role-permission.add-role',compact('role','permissions'));
@@ -46,6 +51,10 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        $count = Role::where('slug','super-admin')->where('id',$id)->count();
+        if($count > 0){
+            return redirect()->back();
+        }
         $data = Role::find($id);
         $data->delete();
         session()->flash('success', 'Role Deleted successfully!');
